backend/requirements.txt
pinecone
python-dotenv
beautifulsoup4 
playwright
fastapi
uvicorn
pandas
numpy
scikit-learn
sentence-transformers
openpyxl 
google-generativeai